improved canvas zooming

+++ main.py
@@ -24,7 +24,7 @@ class Main(Tk):
         self.shape: str = self.shape_options[0]
         self.pixel_size = 50
         self.canvas_size: list[int] = [20, 15]
-        self.scale: int = 100
+        self.scale: float = 100
 
         self.toolbar = Toolbar(self)
         self.main_canvas = MainCanvas(self)
